ZBUG-1948: Fix for the malformed inline styling in
the original message body which contains a lot of unclosed
double quotes which was escaping large chunks of the real
message.

Looking inside the owasp java sanitizer code turns out that
the owasp uses HtmlLexer for the sanitization process which
traverses element by element so before sanitizing the html
it takes the value of the attribute and strips of any of the
vulnerable kinds of stuff based on what policies we have applied.
In this particular case because of the malformed double quotes
which were present inside the tag it thinks of it as the new
attribute has started and intercepts whatever is there as its
value before the next double-quotes.

To fix this issue extended the earlier work of fixing the double
quotes issue addressed in the TSS-18004, but in that particular
issue double-quotes was handled if it was present at the very end
of the tag near to ">" but in this case the the duplicate quotes
were present not at the very end resulting in failing of the issue
as it was addressed in the earlier attempt, to address this issue
added another condition to handle the duplicate quotes if they are
present not at the very end of the tag but in the middle for an
attribute.